disabling unusable formatting support
Right now the server and the client fall out of sync when formatting edits are applied. Need to look into this, probably in a different branch since we require a recently-fixed version of DrRacket's indenter, and we're going to need a version bump when we start making the code that uses it available to users.