Fix indentation after new line at first line of expression
Adding a new line at the end of the first line of an existing expression would
force you back to the start of the line unexpectedly.

For example, with the following:

```racket
(define (bob); <- place cursor here, then press Return
  (+ 1 2))
```

Previously the langserver left the new line unindented. This fixes the
indentation path to instead indent the new line as expected.